use float tensors throughout the computation of the f1score

+++ metrics/f1score.py
@@ -4,7 +4,7 @@ import torch
 def f1score(prediction_logits, targets, epsilon=1e-7):
     predictions = torch.sigmoid(prediction_logits)
 
-    positives = predictions > 0.5
+    positives = (predictions > 0.5).float()
     true_positives = positives * targets
 
     precision = true_positives.sum(dim=1) / (positives.sum(dim=1) + epsilon)
